requirements.txt

huggingface_hub 
transformers 
sentence_transformers 
accelerate bitsandbytes
pinecone-client==2.2.1
python-dotenv==1.0.0
langchain 
